// generated/models/ApiProfileCmsWalletGalleryExcludedAsset.ts
// @ts-nocheck
/**
 * 6529.io API
 * This is the API interface description. Brief terminology overview and an authentication example can be found at <a href=\"https://6529.io/about/api\">https://6529.io/about/api</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class ApiProfileCmsWalletGalleryExcludedAsset {
    'contract': string;
    'token_id': number;
    'owner_wallet': string;
    'reason': ApiProfileCmsWalletGalleryExcludedAssetReasonEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "contract",
            "baseName": "contract",
            "type": "string",
            "format": ""
        },
        {
            "name": "token_id",
            "baseName": "token_id",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "owner_wallet",
            "baseName": "owner_wallet",
            "type": "string",
            "format": ""
        },
        {
            "name": "reason",
            "baseName": "reason",
            "type": "ApiProfileCmsWalletGalleryExcludedAssetReasonEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiProfileCmsWalletGalleryExcludedAsset.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiProfileCmsWalletGalleryExcludedAssetReasonEnum {
    ContractExcluded = 'contract_excluded',
    AssetExcluded = 'asset_excluded'
}