// generated/models/ApiProfileCmsWalletGalleryWallet.ts
// @ts-nocheck
/**
 * 6529.io API
 * This is the API interface description. Brief terminology overview and an authentication example can be found at <a href=\"https://6529.io/about/api\">https://6529.io/about/api</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class ApiProfileCmsWalletGalleryWallet {
    'input': string;
    'address': string | null;
    'ens': string | null;
    'display': string | null;
    'status': ApiProfileCmsWalletGalleryWalletStatusEnum;
    'reason': string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "input",
            "baseName": "input",
            "type": "string",
            "format": ""
        },
        {
            "name": "address",
            "baseName": "address",
            "type": "string",
            "format": ""
        },
        {
            "name": "ens",
            "baseName": "ens",
            "type": "string",
            "format": ""
        },
        {
            "name": "display",
            "baseName": "display",
            "type": "string",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "ApiProfileCmsWalletGalleryWalletStatusEnum",
            "format": ""
        },
        {
            "name": "reason",
            "baseName": "reason",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiProfileCmsWalletGalleryWallet.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiProfileCmsWalletGalleryWalletStatusEnum {
    Resolved = 'resolved',
    Unresolved = 'unresolved'
}